TSK-1884: Move builder and their tests into taskana-test-api
