+++ lib/taskana-core-test/src/test/java/acceptance/task/create/CreateTaskWithSorAccTest.java
@@ -18,17 +18,17 @@ import pro.taskana.task.api.TaskService;
 import pro.taskana.task.api.models.ObjectReference;
 import pro.taskana.task.api.models.Task;
 import pro.taskana.task.internal.ObjectReferenceMapper;
-import pro.taskana.task.internal.builder.ObjectReferenceBuilder;
-import pro.taskana.task.internal.builder.TaskBuilder;
 import pro.taskana.task.internal.models.TaskImpl;
 import pro.taskana.testapi.TaskanaEngineProxy;
 import pro.taskana.testapi.TaskanaInject;
 import pro.taskana.testapi.TaskanaIntegrationTest;
+import pro.taskana.testapi.builder.ObjectReferenceBuilder;
+import pro.taskana.testapi.builder.TaskBuilder;
+import pro.taskana.testapi.builder.WorkbasketAccessItemBuilder;
 import pro.taskana.testapi.security.WithAccessId;
 import pro.taskana.workbasket.api.WorkbasketPermission;
 import pro.taskana.workbasket.api.WorkbasketService;
 import pro.taskana.workbasket.api.models.WorkbasketSummary;
-import pro.taskana.workbasket.internal.builder.WorkbasketAccessItemBuilder;
 
 /**
  * Acceptance test for all "create task" scenarios that involve secondary {@link ObjectReference}.
