+++ lib/taskana-test-api/src/main/java/pro/taskana/testapi/builder/ClassificationTestImpl.java
@@ -1,4 +1,4 @@
-package pro.taskana.classification.internal.builder;
+package pro.taskana.testapi.builder;
 
 import java.time.Instant;
