+++ lib/taskana-test-api/src/main/java/pro/taskana/testapi/builder/TaskAttachmentBuilder.java
@@ -1,4 +1,4 @@
-package pro.taskana.task.internal.builder;
+package pro.taskana.testapi.builder;
 
 import java.time.Instant;
 import java.util.Map;
