+++ lib/taskana-test-api/src/main/java/pro/taskana/testapi/builder/WorkbasketTestImpl.java
@@ -1,4 +1,4 @@
-package pro.taskana.workbasket.internal.builder;
+package pro.taskana.testapi.builder;
 
 import java.time.Instant;
