+++ lib/taskana-test-api/src/test/java/pro/taskana/testapi/tests/DefaultTestEntities.java
@@ -1,11 +1,11 @@
-package acceptance;
+package pro.taskana.testapi.tests;
 
 import java.util.UUID;
 
-import pro.taskana.classification.internal.builder.ClassificationBuilder;
-import pro.taskana.task.internal.builder.ObjectReferenceBuilder;
+import pro.taskana.testapi.builder.ClassificationBuilder;
+import pro.taskana.testapi.builder.ObjectReferenceBuilder;
+import pro.taskana.testapi.builder.WorkbasketBuilder;
 import pro.taskana.workbasket.api.WorkbasketType;
-import pro.taskana.workbasket.internal.builder.WorkbasketBuilder;
 
 public class DefaultTestEntities {
