+++ lib/taskana-test-api/src/test/resources/simplelogger.properties
@@ -1,35 +1,28 @@
 # SLF4J's SimpleLogger configuration file
 # Simple implementation of Logger that sends all enabled log messages, for all defined loggers, to System.err.
-
 # Default logging detail level for all instances of SimpleLogger.
 # Must be one of ("trace", "debug", "info", "warn", or "error").
 # If not specified, defaults to "info".
 org.slf4j.simpleLogger.defaultLogLevel=info
-
 # Logging detail level for a SimpleLogger instance named "xxxxx".
 # Must be one of ("trace", "debug", "info", "warn", or "error").
 # If not specified, the default logging detail level is used.
 org.slf4j.simpleLogger.log.pro.taskana=info
 org.slf4j.simpleLogger.log.org.apache.ibatis=info
-
 # Set to true if you want the current date and time to be included in output messages.
 # Default is false, and will output the number of milliseconds elapsed since startup.
 org.slf4j.simpleLogger.showDateTime=true
-
 # The date and time format to be used in the output messages.
 # The pattern describing the date and time format is the same that is used in java.text.SimpleDateFormat.
 # If the format is not specified or is invalid, the default format is used.
 # The default format is yyyy-MM-dd HH:mm:ss:SSS Z.
 org.slf4j.simpleLogger.dateTimeFormat=yyyy-MM-dd HH:mm:ss:SSS Z
-
 # Set to true if you want to output the current thread name.
 # Defaults to true.
 org.slf4j.simpleLogger.showThreadName=true
-
 # Set to true if you want the Logger instance name to be included in output messages.
 # Defaults to true.
 #org.slf4j.simpleLogger.showLogName=true
-
 # Set to true if you want the last component of the name to be included in output messages.
 # Defaults to false.
 org.slf4j.simpleLogger.showShortLogName=false
